+++ src/main/java/com/gitlab/juli220620/service/GameSystemService.java
@@ -33,11 +33,12 @@ public class GameSystemService {
     /**
      * <p>Add system to user on 1 level and returns map of costs by currency.
      * <b>It does not spend any user currencies by itself</b></p>
-     * @param user - recipient of system
+     *
+     * @param user     - recipient of system
      * @param systemId - system to receive
      * @return Map of upgrade costs by currency to subtract from user funds
      */
-    public Map<String, Integer> addSystem(UserEntity user, String systemId) {
+    public Map<String, Long> addSystem(UserEntity user, String systemId) {
         GameSystemEntity system = gameSystemRepo.findById(systemId).orElseThrow();
 
         user.getWorkingSystems().stream()
@@ -60,18 +61,19 @@ public class GameSystemService {
     /**
      * <p>Upgrades user active system by adding 1 level to it and returns map of costs for upgrade by currency.
      * <b>It does not spend any user currencies by itself</b></p>
-     * @param user - recipient of upgrade
+     *
+     * @param user     - recipient of upgrade
      * @param systemId - system to upgrade for recipient
      * @return Map of upgrade costs by currency to subtract from user funds
      */
-    public Map<String, Integer> upgradeSystem(UserEntity user, String systemId) {
+    public Map<String, Long> upgradeSystem(UserEntity user, String systemId) {
         UserGameSystemEntity userSystem = user.getWorkingSystems().stream()
                 .filter(it -> it.getId().getSystemId().contentEquals(systemId))
                 .findFirst()
                 .orElseThrow(() -> new RuntimeException("No such system"));
 
         GameSystem target = systems.get(systemId);
-        Map<String, Integer> costs = Optional.ofNullable(target.getCost().get(userSystem.getSystemLevel() + 1))
+        Map<String, Long> costs = Optional.ofNullable(target.getCost().get(userSystem.getSystemLevel() + 1))
                 .orElseThrow(() -> new RuntimeException("Max level"));
 
         userSystem.setSystemLevel(userSystem.getSystemLevel() + 1);
