+++ src/main/java/com/gitlab/juli220620/web/controllers/dict/BaseFlowerDto.java
@@ -25,7 +25,7 @@ public class BaseFlowerDto {
     private Integer price;
     private Integer maxCycles;
 
-    private Map<String, Integer> harvest;
+    private Map<String, Long> harvest;
 
     private List<HarvestBonusDto> harvestBonuses;
